fine progetto miniCRM

+++ app/Http/Controllers/ClientController.php
@@ -68,10 +68,11 @@ class ClientController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Client $client)
+    public function destroy(Client $client): RedirectResponse
     {
-        /* Gate::authorize(PermissionEnum::DELETE_CLIENTS->value);
- */
+        Gate::authorize(PermissionEnum::DELETE_CLIENTS->value);
+
+        $client->projects()->delete();
         $client->delete();
 
         return redirect()->route('clients.index');
